Move resource manager objects to Base project
